Enabling all cpp files in Examples for compilation.

+++ Libraries/fermiqcd_hmc.h
@@ -154,13 +154,13 @@ public:
       numgen = g.ngenerators;
       to_V = to_U; // default      
       lambda.resize(g.ngenerators);
-      for(int a=0; a<lambda.size(); a++) lambda[a] = g.lambda[a];
+      for(size_t a=0; a<lambda.size(); a++) lambda[a] = g.lambda[a];
     } else if(coeff["representation"] == SYMMETRIC) {
       SU_Generators g(U.nc);
       numgen = g.ngenerators;
       int dimrep = U.nc*(U.nc+1)/2;
       lambda.resize(g.ngenerators);
-      for(int a=0; a<lambda.size(); a++) lambda[a] = g.lambda[a];
+      for(size_t a=0; a<lambda.size(); a++) lambda[a] = g.lambda[a];
       to_V = new GaugeClass(U.lattice(), dimrep);
       mdp_matrix tmp(U.nc, U.nc);
       S.resize(g.ngenerators);
@@ -186,7 +186,7 @@ public:
       SO_Generators g(4);
       numgen = g.ngenerators;
       lambda.resize(g.ngenerators);
-      for(int a=0; a<lambda.size(); a++) lambda[a] = g.lambda[a];
+      for(size_t a=0; a<lambda.size(); a++) lambda[a] = g.lambda[a];
       to_V = to_U;
     }    
     /*
